Fix chainable methods to return proxy object
Fix methods that are chainable to return the PheanstalkProxy object instead of the Pheanstalk_Pheanstalk object. This allows code to always use the proxy even when methods are changed. Without it code like:
$pheanstalk->useTube('testTube')->put(json_encode($jobData));
would not trigger any event subscribers because put() would be executed directly on the Pheanstalk_Pheanstalk object.